Change generation of protojs file to es5 (commonjs)
Use protobufjs to generate an es5 output instead of es6. This leads
to easier usage, as the output does not need transpilation if used
in a fully es5 environment.
