+++ packages/bs-protobuf/src/compiler/gen-types/emit-types.js
@@ -15,9 +15,7 @@ function emitPrologue(stream, resolver) {
   stream.write(`\
 // Generated by protores from @reproto/bs-protobuf, PLEASE EDIT WITH CARE
 type protoJs = unit
-`);
-  emitProtoModuleDirective(stream, resolver);
-  stream.write(`@val external protoJs: protoJs = "default"
+@module() @val external protoJs: protoJs = "${resolver.options.protoJsPath}"
 type serviceRoot
 @module("@reproto/bs-protobuf") external _createServiceRoot: _ => serviceRoot = "createServiceRoot"
 let createServiceRoot = () => _createServiceRoot(protoJs)
@@ -26,15 +24,16 @@ type rec protoStreamCallback<'res> = ${bsProtobufPackage}.MethodWrapper.protoStr
 let methodWrapper = ${bsProtobufPackage}.MethodWrapper.methodWrapper
 module ${specialModuleTypes} = {
 `);
+  // Noting the options for es5 and es6 generated protobufjs outputs.
+  // es5:
+  // @module() @val external protoJs: protoJs = "${resolver.options.protoJsPath}"
+  // es6:
+  // @module("${resolver.options.protoJsPath}") @val external protoJs: protoJs = "default"
 }
 
 function emitBetweenPasses(stream, resolver) {
   stream.write(`}
 `);
-}
-
-function emitProtoModuleDirective(stream, resolver) {
-  stream.write(`@module("${resolver.options.protoJsPath}") `);
 }
 
 function emitEpilogue(stream, resolver) {}
